variable hover menu for delete/edit, rm control node type
refactored variable nodes structure to directly render control within the variable after the label, hint. allows for easier control of attributes and removes unneccessary nesting
